requirements.txt
Flask>=2.2.0
Flask-Cors>=3.0.0
numpy>=1.23.0
opencv-python-headless>=4.6.0
Pillow>=9.5.0
pypotrace>=0.3
gunicorn>=21.0.0
